Meta-tag fix for iOS; audio controls in navbar
I was having issues with iOS not playing audio tracks when clicked.  It
worked fine in a browser, iOS errored out.  The added meta tag disables
the "this looks like a phone number, I should treat it as a phone
number" feature which was causing the frequency to be tagged as a phone
number.  After that, iOS plays basically normally.

The other change was to put the audio controls in a navbar glued to the
bottom of the screen. Generally you move down the playlist pretty quick
and if you want to stop playing, you had to scroll all the way to the
top.  This keeps the controls always in reach.