Correct TGID lookup in audioplayer.php
So this update corrects one issue I've found with the last update to support multiple systems.  While the file lookup were working fine, the TGID updates weren't separating by system.  So if multiple systems are in place, and they had overlapping decimal talkgroups whichever system was loaded last got the correct alphatag.  The major part of this update separates $TGS by $SHORTNAME.  $TGS becomes a multi-dimensional array of [$SHORTNAME][TGID] and the lookup is changed to match.

There are a few other updates to the file as a result of html5 validation:

* lang is set in the <html>
* The <style> element moves into the <head>

And finally there are a couple of cosmetic updates which are completely optional:

* There is now a "jump to bottom" button which scrolls the page to the bottom (the most recent entries).
* I changed the bootstrap colors to primary which seems to work better when flipping themes