Removing set_max_output_buffer
Setting the max_output_buffer seemed to cause a lot of problems. The buffer gets sized dynamically depending on the system config, so 4096 might be too low for some systems. It was added in to try and keep the buffers as small as possible to prevent audio from call sticking in the buffer and ending up in another call.
Removing ->set_max_output_buffer(4096);
#319