Encrypted E MODE in Talkgroups.csv file will never record that talkgroup.
Keep in mind that `Talkgroup::Talkgroup` mode is a single char, as such you can only set `E` and not `DE`. The only there place this is checked is for `A` as in analog, so it shouldn't break anything else. Also fine with me as encrypted also implies that it must be digital.