Fix a filename issue for analog recorders
when a source changes within 1 second, then the filenames will be the same leading to the file being included twice in the list and doubling up. The change ensures that filenames are unique. It also will not start a new transmission if the source ID change is from 0 to a new source ID