Update main.cc - fix for segfault on conventional systems
In a conventional system the sys object isn't initialized since a control channel message is never received, leading to a segfault when the clear_stale_talkgroup_patches() function gets called and gets passed an empty System object.  This check prevents calling the function unless the system type is known as p25, which means by definition that it's not empty.