Update vendored Eigen to version 3.4
The current version of Eigen that is vendored in lib/lfsr
is 5 to 6 years old (between commits 6bceebfab to 68e8f2b83).
Additionally, increasing the C++ version results in warnings about
std::unary_negate and std::binary_negate, which are deprecated in
C++17 and removed in C++20. Eigen 3.3 contains a fix for this.