add -D_GLIBCXX_ASSERTIONS to CXXFLAGS
Many linux distributions have switched this on by default, but TR does
not have it on by default and so bugs are slipping through and hitting
users.  Turn on bounds checking by default so that bugs are found
quicker.

This WILL break users.  I'm sure of it.  It will cause program exit when
running broken code, example being https://github.com/robotastic/trunk-recorder/issues/783
The performance penalty of this is minimal (none on modern distros where
this is already teh default) and the bounds checking causing failure is
better than simply running objectively wrong code.